add: cd command and related files
